Remove parse_modules module
